Extract image handling for reuse in pattern
