+++ shoes-swt/lib/shoes/swt.rb
@@ -59,6 +59,7 @@ class Shoes
       require 'shoes/swt/common/clickable'
       require 'shoes/swt/common/container'
       require 'shoes/swt/common/fill'
+      require 'shoes/swt/common/image_handling'
       require 'shoes/swt/common/resource'
       require 'shoes/swt/common/painter'
       require 'shoes/swt/common/painter_updates_position'
