+++ shoes-swt/lib/shoes/swt/image_pattern.rb
@@ -3,6 +3,7 @@ class Shoes
   module Swt
     class ImagePattern
       include Common::Remove
+      include Common::ImageHandling
 
       def initialize(dsl)
         @dsl = dsl
@@ -16,8 +17,11 @@ class Shoes
       # Since colors are bound up (at least in specs) with image patterns,
       # we can't safely touch images during initialize, so lazily load them.
       def pattern
-        @image   ||= ::Swt::Image.new(Shoes.display, @dsl.path)
+        @image   ||= ::Swt::Image.new(Shoes.display, load_file_image_data(@dsl.path))
         @pattern ||= ::Swt::Pattern.new(Shoes.display, @image)
+        cleanup_temporary_files
+
+        @pattern
       end
 
       def apply_as_fill(gc, _dsl)
